Remove most of the `/v1` endpoints from core
